--HG--
branch : pilfer

+++ lib/python/cs/cmdutils.py
@@ -9,6 +9,7 @@ import os
 import sys
 import io
 import subprocess
+from cs.pfx import Pfx
 
 def run(argv, trace=False):
   ''' Run a command. Optionally trace invocation. Return result of subprocess.call.
@@ -23,9 +24,10 @@ def run(argv, trace=False):
     print(*pargv, file=tracefp)
   return subprocess.call(argv)
 
-def pipefrom(argv, trace=False, **kw):
+def pipefrom(argv, trace=False, binary=False, keep_stdin=False, **kw):
   ''' Pipe text from a command. Optionally trace invocation. Return the Popen object with .stdout decoded as text.
       `argv`: the command argument list
+      `binary`: if true (default false) return the binary stdout instead of a text wrapper
       `trace`: Default False. If True, recite invocation to stderr.
         Otherwise presume a stream to which to recite the invocation.
       The command's stdin is attached to the null device.
@@ -33,16 +35,23 @@ def pipefrom(argv, trace=False, **kw):
   '''
   if trace:
     tracefp = sys.stderr if trace is True else trace
-    pargv = ['+'] + argv + ['|']
+    pargv = ['+'] + list(argv) + ['|']
     print(*pargv, file=tracefp)
-  sp_devnull = getattr(subprocess, 'DEVNULL', None)
-  if sp_devnull is None:
-    devnull = open(os.devnull, 'wb')
+  popen_kw = {}
+  if not keep_stdin:
+    sp_devnull = getattr(subprocess, 'DEVNULL', None)
+    if sp_devnull is None:
+      devnull = open(os.devnull, 'wb')
+    else:
+      devnull = sp_devnull
+    popen_kw['stdin'] = devnull
+  P = subprocess.Popen(argv, stdout=subprocess.PIPE, **popen_kw)
+  if binary:
+    if kw:
+      raise ValueError("binary mode: extra keyword arguments not supported: %r", kw)
   else:
-    devnull = sp_devnull
-  P = subprocess.Popen(argv, stdin=devnull, stdout=subprocess.PIPE)
-  P.stdout = io.TextIOWrapper(P.stdout, **kw)
-  if sp_devnull is None:
+    P.stdout = io.TextIOWrapper(P.stdout, **kw)
+  if not keep_stdin and sp_devnull is None:
     devnull.close()
   return P
 
@@ -60,3 +69,24 @@ def pipeto(argv, trace=False, **kw):
   P = subprocess.Popen(argv, stdin=subprocess.PIPE)
   P.stdin = io.TextIOWrapper(P.stdin, **kw)
   return P
+
+def docmd(dofunc):
+  ''' Decorator for Cmd subclass methods.
+  '''
+  def wrapped(self, *a, **kw):
+    funcname = dofunc.__name__
+    if not funcname.startswith('do_'):
+      raise ValueError("function does not start with 'do_': %s" % (funcname,))
+    argv0 = funcname[3:]
+    with Pfx(argv0):
+      try:
+        return dofunc(self, *a, **kw)
+      except GetoptError as e:
+        warning("%s", e)
+        self.do_help(argv0)
+        return None
+      except Exception as e:
+        exception("%s", e)
+        return None
+  wrapped.__doc__ = dofunc.__doc__
+  return wrapped
